requirements.txt
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
drf-nested-routers==0.90.2
pkg-resources==0.0.0
pytz==2018.4
